removed some warnings
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@5638 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rwlibs/algorithms/StructuredLineModel.cpp
@@ -153,7 +153,7 @@ double StructuredLineModel::testInterval(const std::vector<rw::math::Vector3D<> 
 
 double StructuredLineModel::refit(const std::vector<rw::math::Vector3D<> >& samples)
 {
-	if (samples.size() < getMinReqData()) {
+	if ((int)samples.size() < getMinReqData()) {
 		RW_THROW("Too few samples to refit structured line model.");
 		return 0.0;
 	}
